chore: move src & tests from local to cli
