+++ packages/cli/bin/bats-test-runner.js
