+++ packages/cli/test/acceptance/run.bats
