+++ packages/cli/test/acceptance/start.bats
