+++ packages/cli/test/acceptance/version.bats
